requirements.txt
numpy==1.24.3
pandas==2.0.2
prefect==2.10.13
prefect-gcp==0.4.2
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
pyspark==3.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.1.0
readchar==4.0.5
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
stack-data==0.6.2
starlette==0.27.0
text-unidecode==1.3
toml==0.10.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.6
websocket-client==1.5.3
websockets==11.0.3
yarl==1.9.2
